pyproject.toml
[tool.poetry]
name = "SorTgBot"
version = "0.1.0"
description = ""
authors = ["Rasul Taimaskhanov <81189712+taimaskhanov11@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
aiogram = "^2.19"
loguru = "^0.6.0"
tortoise-orm = { extras = ["asyncpg"], version = "^0.19.0" }
PyYAML = "^6.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.1.13"]
build-backend = "poetry.core.masonry.api"
